moving files into the lib folder
